### 2021-05-27 8246009
CompatHelper: add new compat entry for "StaticArrays" at version "1.2" for package docs
