+++ docs/Project.toml
@@ -18,3 +18,4 @@ Documenter = "0.26"
 LaTeXStrings = "1.2"
 OrdinaryDiffEq = "5.50"
 Plots = "1.9"
+StaticArrays = "1.2"
