+++ paper.bib
@@ -199,17 +199,20 @@
 }
 
 
-@online{schlottkelakemper2021purely,
+@article{schlottkelakemper2021purely,
   title={A purely hyperbolic discontinuous {G}alerkin approach for
          self-gravitating gas dynamics},
   author={Schlottke-Lakemper, Michael and Winters, Andrew R and
           Ranocha, Hendrik and Gassner, Gregor J},
+  journal={Journal of Computational Physics},
+  pages={110467},
   year={2021},
   month={05},
+  publisher={Elsevier},
+  doi={10.1016/j.jcp.2021.110467},
   eprint={2008.10593},
   eprinttype={arXiv},
-  eprintclass={math.NA},
-  note={Accepted in Journal of Computational Physics}
+  eprintclass={math.NA}
 }
 @misc{schlottkelakemper2020trixi,
   title={{T}rixi.jl: A tree-based numerical simulation framework
@@ -279,7 +282,7 @@
   number={52},
   pages={2520},
   year={2020},
-  doi={0.21105/joss.02520}
+  doi={10.21105/joss.02520}
 }
 
 @article{mattsson2014optimal,
