### 2021-05-30 84add11
specialize mul! for arrays of SVector{1}
