+++ test/special_matrix_types.jl
@@ -80,6 +80,23 @@ for T in (Float32, Float64)
   fill!(du_soa, zero(eltype(du_soa)))
   mul!(du_soa, D, u_soa, α, β)
   @test du_soa ≈ du_aos
+
+
+  # `SVector` with only one element
+  u_scalar = randn(T, size(D, 1)); du_scalar = similar(u_scalar)
+  u_vector = reinterpret(SVector{1,T}, u_scalar); du_vector = similar(u_vector)
+
+  mul!(du_scalar, D, u_scalar)
+  mul!(du_vector, D, u_vector)
+  @test du_scalar ≈ reinterpret(T, du_vector)
+
+  mul!(du_scalar, D, u_scalar, α)
+  mul!(du_vector, D, u_vector, α)
+  @test du_scalar ≈ reinterpret(T, du_vector)
+
+  mul!(du_scalar, D, u_scalar, α, β)
+  mul!(du_vector, D, u_vector, α, β)
+  @test du_scalar ≈ reinterpret(T, du_vector)
 end
 
 end # module
