### 2021-05-30 1792471
get real type of SBP operators
