### 2021-05-31 4f0867b
DiffEqCallbacks via Requires; closes #98
