+++ Project.toml
@@ -6,7 +6,6 @@ version = "0.5.3-pre"
 [deps]
 ArgCheck = "dce04be8-c92d-5529-be00-80e4d2c0e197"
 DiffEqBase = "2b5f629d-d688-5b77-993f-72d75c75574e"
-DiffEqCallbacks = "459566f4-90b8-5000-8ac3-15dfb0a30def"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 InteractiveUtils = "b77e0a4c-d291-57a0-90e8-8db25a27a240"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
@@ -23,7 +22,6 @@ Unrolled = "9602ed7d-8fef-5bc8-8597-8f21381861e8"
 [compat]
 ArgCheck = "1, 2.0"
 DiffEqBase = "6"
-DiffEqCallbacks = "2.6"
 FFTW = "1"
 LoopVectorization = "0.12.22"
 PolynomialBases = "0.4.5"
