### 2021-05-31 2a6e1a3
make tests a bit more verbose
