### 2021-05-31 4ae9a09
unify convolve_interior_coefficients for periodic and non-periodic operators; closes #97
