+++ src/SBP_operators.jl
@@ -86,7 +86,7 @@ function mul!(dest::AbstractVector, coefficients::DerivativeCoefficients, u::Abs
     end
 
     convolve_boundary_coefficients!(dest, left_boundary, right_boundary, u, α, β)
-    convolve_interior_coefficients!(dest, lower_coef, central_coef, upper_coef, u, α, β, length(left_boundary), length(right_boundary), parallel)
+    convolve_interior_coefficients!(dest, lower_coef, central_coef, upper_coef, u, α, β, static_length(left_boundary), static_length(right_boundary), parallel)
 end
 
 # Compute `α*D*u` and store the result in `dest`.
@@ -100,7 +100,7 @@ function mul!(dest::AbstractVector, coefficients::DerivativeCoefficients, u::Abs
     end
 
     convolve_boundary_coefficients!(dest, left_boundary, right_boundary, u, α)
-    convolve_interior_coefficients!(dest, lower_coef, central_coef, upper_coef, u, α, length(left_boundary), length(right_boundary), parallel)
+    convolve_interior_coefficients!(dest, lower_coef, central_coef, upper_coef, u, α, static_length(left_boundary), static_length(right_boundary), parallel)
 end
 
 
@@ -198,7 +198,11 @@ end
 end
 
 
-@generated function convolve_interior_coefficients!(dest::AbstractVector, lower_coef::SVector{LowerOffset}, central_coef, upper_coef::SVector{UpperOffset}, u::AbstractVector, α, β, left_boundary_width, right_boundary_width, parallel) where {LowerOffset, UpperOffset}
+@generated function convolve_interior_coefficients!(
+        dest::AbstractVector,
+        lower_coef::SVector{LowerOffset}, central_coef, upper_coef::SVector{UpperOffset},
+        u::AbstractVector, α, β,
+        ::StaticInt{left_boundary_width}, ::StaticInt{right_boundary_width}, parallel) where {LowerOffset, UpperOffset, left_boundary_width, right_boundary_width}
     if LowerOffset > 0
         ex = :( lower_coef[$LowerOffset]*u[i-$LowerOffset] )
         for j in LowerOffset-1:-1:1
@@ -216,6 +220,13 @@ end
         quote
             Base.@_inline_meta
             @tturbo for i in (left_boundary_width+1):(length(dest)-right_boundary_width)
+                dest[i] = β*dest[i] + α*$ex
+            end
+        end
+    elseif parallel <: Val{:plain}
+        quote
+            Base.@_inline_meta
+            @inbounds for i in (left_boundary_width+1):(length(dest)-right_boundary_width)
                 dest[i] = β*dest[i] + α*$ex
             end
         end
@@ -229,7 +240,7 @@ end
     end
 end
 
-@generated function convolve_interior_coefficients!(dest::AbstractVector, lower_coef::SVector{LowerOffset}, central_coef, upper_coef::SVector{UpperOffset}, u::AbstractVector, α, left_boundary_width, right_boundary_width, parallel) where {LowerOffset, UpperOffset}
+@generated function convolve_interior_coefficients!(dest::AbstractVector, lower_coef::SVector{LowerOffset}, central_coef, upper_coef::SVector{UpperOffset}, u::AbstractVector, α, ::StaticInt{left_boundary_width}, ::StaticInt{right_boundary_width}, parallel) where {LowerOffset, UpperOffset, left_boundary_width, right_boundary_width}
     if LowerOffset > 0
         ex = :( lower_coef[$LowerOffset]*u[i-$LowerOffset] )
         for j in LowerOffset-1:-1:1
@@ -246,14 +257,21 @@ end
     if parallel <: Val{:threads}
         quote
             Base.@_inline_meta
-            @tturbo for i in (left_boundary_width+1):(length(dest)-right_boundary_width)
[…]
