### 2021-05-31 da23e56
respect comapct setting of IO for printing; closes #99
