+++ test/periodic_operators_test.jl
@@ -23,7 +23,9 @@ let T=Float32
     derivative_order = 1
     accuracy_order = 2
     D = periodic_central_derivative_operator(derivative_order, accuracy_order, xmin, xmax, N)
-    println(devnull, D)
+    for compact in (true, false)
+        show(IOContext(devnull, :compact=>false), D)
+    end
     @test SummationByPartsOperators.derivative_order(D) == derivative_order
     @test SummationByPartsOperators.accuracy_order(D) == accuracy_order
     @test issymmetric(D) == false
@@ -36,7 +38,9 @@ let T=Float32
 
     accuracy_order = 4
     D = periodic_central_derivative_operator(derivative_order, accuracy_order, xmin, xmax, N)
-    println(devnull, D)
+    for compact in (true, false)
+        show(IOContext(devnull, :compact=>false), D)
+    end
     @test SummationByPartsOperators.derivative_order(D) == derivative_order
     @test SummationByPartsOperators.accuracy_order(D) == accuracy_order
     @test issymmetric(D) == false
@@ -53,7 +57,9 @@ let T=Float32
 
     accuracy_order = 6
     D = periodic_central_derivative_operator(derivative_order, accuracy_order, xmin, xmax, N)
-    println(devnull, D)
+    for compact in (true, false)
+        show(IOContext(devnull, :compact=>false), D)
+    end
     @test SummationByPartsOperators.derivative_order(D) == derivative_order
     @test SummationByPartsOperators.accuracy_order(D) == accuracy_order
     @test issymmetric(D) == false
@@ -76,7 +82,9 @@ let T=Float32
     derivative_order = 2
     accuracy_order = 2
     D = periodic_central_derivative_operator(derivative_order, accuracy_order, xmin, xmax, N)
-    println(devnull, D)
+    for compact in (true, false)
+        show(IOContext(devnull, :compact=>false), D)
+    end
     @test SummationByPartsOperators.derivative_order(D) == derivative_order
     @test SummationByPartsOperators.accuracy_order(D) == accuracy_order
     @test issymmetric(D) == true
@@ -93,7 +101,9 @@ let T=Float32
 
     accuracy_order = 4
     D = periodic_central_derivative_operator(derivative_order, accuracy_order, xmin, xmax, N)
-    println(devnull, D)
+    for compact in (true, false)
+        show(IOContext(devnull, :compact=>false), D)
+    end
     @test SummationByPartsOperators.derivative_order(D) == derivative_order
     @test SummationByPartsOperators.accuracy_order(D) == accuracy_order
     @test issymmetric(D) == true
@@ -112,7 +122,9 @@ let T=Float32
 
     accuracy_order = 6
     D = periodic_central_derivative_operator(derivative_order, accuracy_order, xmin, xmax, N)
-    println(devnull, D)
+    for compact in (true, false)
+        show(IOContext(devnull, :compact=>false), D)
+    end
     @test SummationByPartsOperators.derivative_order(D) == derivative_order
     @test SummationByPartsOperators.accuracy_order(D) == accuracy_order
     @test issymmetric(D) == true
@@ -133,7 +145,9 @@ let T=Float32
     derivative_order = 3
     accuracy_order = 2
     D = periodic_central_derivative_operator(derivative_order, accuracy_order, xmin, xmax, N)
-    println(devnull, D)
+    for compact in (true, false)
+        show(IOContext(devnull, :compact=>false), D)
+    end
     @test SummationByPartsOperators.derivative_order(D) == derivative_order
     @test SummationByPartsOperators.accuracy_order(D) == accuracy_order
     @test issymmetric(D) == true # because this operator is zero!
@@ -144,7 +158,9 @@ let T=Float32
 
[…]
