### 2021-05-31 4ad4c6b
fix doctests
