### 2021-05-31 01a0b6f
introduce AbstractExecutionMode and deprecate parallel in favor of mode; closes #101
