+++ src/SBP_coefficients/MattssonAlmquistCarpenter2014Optimal.jl
@@ -109,7 +109,8 @@ function construct_grid(::MattssonAlmquistCarpenter2014Optimal, accuracy_order, 
 end
 
 
-function first_derivative_coefficients(source::MattssonAlmquistCarpenter2014Optimal, order::Int, T=Float64, parallel=Val{:serial}())
+function first_derivative_coefficients(source::MattssonAlmquistCarpenter2014Optimal,
+                                       order::Int, T=Float64, mode=FastMode())
     if order == 2
         left_boundary = (
             # d1
@@ -140,7 +141,7 @@ function first_derivative_coefficients(source::MattssonAlmquistCarpenter2014Opti
         DerivativeCoefficients(left_boundary, right_boundary,
                                 left_boundary_derivatives, right_boundary_derivatives,
                                 lower_coef, central_coef, upper_coef,
-                                left_weights, right_weights, parallel, 1, order, source)
+                                left_weights, right_weights, mode, 1, order, source)
     elseif order == 4
         left_boundary = (
             # d1
@@ -193,7 +194,7 @@ function first_derivative_coefficients(source::MattssonAlmquistCarpenter2014Opti
         DerivativeCoefficients(left_boundary, right_boundary,
                                 left_boundary_derivatives, right_boundary_derivatives,
                                 lower_coef, central_coef, upper_coef,
-                                left_weights, right_weights, parallel, 1, order, source)
+                                left_weights, right_weights, mode, 1, order, source)
     elseif order == 6
         left_boundary = (
             # d1
@@ -277,7 +278,7 @@ function first_derivative_coefficients(source::MattssonAlmquistCarpenter2014Opti
         DerivativeCoefficients(left_boundary, right_boundary,
                                 left_boundary_derivatives, right_boundary_derivatives,
                                 lower_coef, central_coef, upper_coef,
-                                left_weights, right_weights, parallel, 1, order, source)
+                                left_weights, right_weights, mode, 1, order, source)
     #= FIXME: There seem to be some errors in the coefficients...
     elseif order == 8
         left_boundary = (
@@ -383,7 +384,7 @@ function first_derivative_coefficients(source::MattssonAlmquistCarpenter2014Opti
         DerivativeCoefficients(left_boundary, right_boundary,
                                 left_boundary_derivatives, right_boundary_derivatives,
                                 lower_coef, central_coef, upper_coef,
-                                left_weights, right_weights, parallel, 1, order, source)
+                                left_weights, right_weights, mode, 1, order, source)
     =#
     else
         throw(ArgumentError("Order $order not implemented/derived."))
