+++ src/SBP_operators.jl
@@ -5,7 +5,7 @@
 The coefficients of a derivative operator on a nonperiodic grid.
 """
 struct DerivativeCoefficients{T,LeftBoundary,RightBoundary,LeftBoundaryDerivatives,RightBoundaryDerivatives,
-                              LowerOffset,UpperOffset,LeftWidth,RightWidth,Parallel,SourceOfCoefficients} <: AbstractDerivativeCoefficients{T}
+                              LowerOffset,UpperOffset,LeftWidth,RightWidth,ExecutionMode,SourceOfCoefficients} <: AbstractDerivativeCoefficients{T}
     # coefficients defining the operator and its action
     left_boundary::LeftBoundary
     right_boundary::RightBoundary
@@ -16,7 +16,7 @@ struct DerivativeCoefficients{T,LeftBoundary,RightBoundary,LeftBoundaryDerivativ
     upper_coef::SVector{UpperOffset, T}
     left_weights::SVector{LeftWidth, T}
     right_weights::SVector{RightWidth, T}
-    parallel::Parallel
+    mode::ExecutionMode
     # corresponding orders etc.
     derivative_order::Int
     accuracy_order  ::Int
@@ -28,9 +28,9 @@ struct DerivativeCoefficients{T,LeftBoundary,RightBoundary,LeftBoundaryDerivativ
                                     right_boundary_derivatives::RightBoundaryDerivatives,
                                     lower_coef::SVector{LowerOffset, T}, central_coef::T, upper_coef::SVector{UpperOffset, T},
                                     left_weights::SVector{LeftWidth, T}, right_weights::SVector{RightWidth, T},
-                                    parallel::Parallel, derivative_order::Int, accuracy_order::Int,
+                                    mode::ExecutionMode, derivative_order::Int, accuracy_order::Int,
                                     source_of_coefficients::SourceOfCoefficients) where {T,LeftBoundary,RightBoundary,LeftBoundaryDerivatives,RightBoundaryDerivatives,
-                                                                                        LowerOffset,UpperOffset,LeftWidth,RightWidth,Parallel,SourceOfCoefficients}
+                                                                                        LowerOffset,UpperOffset,LeftWidth,RightWidth,ExecutionMode,SourceOfCoefficients}
         @argcheck length(left_boundary) == LeftWidth
         @argcheck length(right_boundary) == RightWidth
         @argcheck length(left_boundary_derivatives) == length(right_boundary_derivatives)
@@ -45,10 +45,10 @@ struct DerivativeCoefficients{T,LeftBoundary,RightBoundary,LeftBoundaryDerivativ
         if derivative_order-1 != length(left_boundary_derivatives)
             warn("Derivative coefficients of degree $derivative_order should provide $(derivative_order-1) boundary derivatives.")
         end
-        new{T,LeftBoundary,RightBoundary,LeftBoundaryDerivatives,RightBoundaryDerivatives,LowerOffset,UpperOffset,LeftWidth,RightWidth,Parallel,SourceOfCoefficients}(
+        new{T,LeftBoundary,RightBoundary,LeftBoundaryDerivatives,RightBoundaryDerivatives,LowerOffset,UpperOffset,LeftWidth,RightWidth,ExecutionMode,SourceOfCoefficients}(
             left_boundary, right_boundary, left_boundary_derivatives, right_boundary_derivatives,
             lower_coef, central_coef, upper_coef, left_weights, right_weights,
-            parallel, derivative_order, accuracy_order, symmetric, source_of_coefficients)
+            mode, derivative_order, accuracy_order, symmetric, source_of_coefficients)
[…]
