+++ src/SBP_coefficients/MattssonAlmquistVanDerWeide2018Accurate.jl
@@ -63,7 +63,8 @@ function construct_grid(::MattssonAlmquistVanDerWeide2018Accurate, accuracy_orde
 end
 
 
-function first_derivative_coefficients(source::MattssonAlmquistVanDerWeide2018Accurate, order::Int, T=Float64, parallel=Val{:serial}())
+function first_derivative_coefficients(source::MattssonAlmquistVanDerWeide2018Accurate,
+                                       order::Int, T=Float64, mode=FastMode())
   if order == 4
     left_boundary = (
         # d1
@@ -105,7 +106,7 @@ function first_derivative_coefficients(source::MattssonAlmquistVanDerWeide2018Ac
     DerivativeCoefficients(left_boundary, right_boundary,
                             left_boundary_derivatives, right_boundary_derivatives,
                             lower_coef, central_coef, upper_coef,
-                            left_weights, right_weights, parallel, 1, order, source)
+                            left_weights, right_weights, mode, 1, order, source)
   elseif order == 6
     left_boundary = (
         # d1
@@ -174,7 +175,7 @@ function first_derivative_coefficients(source::MattssonAlmquistVanDerWeide2018Ac
     DerivativeCoefficients(left_boundary, right_boundary,
                             left_boundary_derivatives, right_boundary_derivatives,
                             lower_coef, central_coef, upper_coef,
-                            left_weights, right_weights, parallel, 1, order, source)
+                            left_weights, right_weights, mode, 1, order, source)
   elseif order == 8
     left_boundary = (
         # d1
@@ -279,7 +280,7 @@ function first_derivative_coefficients(source::MattssonAlmquistVanDerWeide2018Ac
     DerivativeCoefficients(left_boundary, right_boundary,
                             left_boundary_derivatives, right_boundary_derivatives,
                             lower_coef, central_coef, upper_coef,
-                            left_weights, right_weights, parallel, 1, order, source)
+                            left_weights, right_weights, mode, 1, order, source)
   # elseif order == 8
   #   left_boundary = (
   #       # d1
@@ -352,7 +353,7 @@ function first_derivative_coefficients(source::MattssonAlmquistVanDerWeide2018Ac
   #   DerivativeCoefficients(left_boundary, right_boundary,
   #                           left_boundary_derivatives, right_boundary_derivatives,
   #                           lower_coef, central_coef, upper_coef,
-  #                           left_weights, right_weights, parallel, 1, order, source)
+  #                           left_weights, right_weights, mode, 1, order, source)
   else
     throw(ArgumentError("Order $order not implemented/derived."))
   end

[…]
