### 2021-06-01 0e66f3e
add compat entry for ArrayInterface
