+++ Project.toml
@@ -22,6 +22,7 @@ Unrolled = "9602ed7d-8fef-5bc8-8597-8f21381861e8"
 
 [compat]
 ArgCheck = "1, 2.0"
+ArrayInterface = "3.1"
 DiffEqBase = "6"
 FFTW = "1"
 LoopVectorization = "0.12.22"
