### 2021-06-03 5d04728
update compat bounds for ArrayInterface.jl
