### 2021-06-03 572f2fe
set development version to v0.5.5-pre [skip ci]
