### 2021-06-29 8830598
set development version to v0.5.6-pre
