### 2021-06-29 eed32b3
cancel redundant CI builds automatically
