+++ .github/workflows/ci.yml
@@ -36,6 +36,11 @@ on:
       - 'utils/**'
   workflow_dispatch:
 
+# Cancel redundant CI tests automatically
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
+
 jobs:
   test:
     if: "!contains(github.event.head_commit.message, 'skip ci')"
