### 2021-07-05 #3454 @whedon
[REVIEW]: SummationByPartsOperators.jl: A Julia library of provably stable discretization techniques with mimetic properties
**Submitting author:** @ranocha (<a href="http://orcid.org/0000-0002-3456-2277">Hendrik Ranocha</a>)
**Repository:** <a href="https://github.com/ranocha/SummationByPartsOperators.jl" target ="_blank">https://github.com/ranocha/SummationByPartsOperators.jl</a>
**Version:** v0.5.7
