+++ versions.js
@@ -1,0 +1,8 @@
+var DOC_VERSIONS = [
+  "stable",
+  "v0.5",
+  "v0.4",
+  "dev",
+];
+var DOCUMENTER_NEWEST = "v0.5.5";
+var DOCUMENTER_STABLE = "stable";
