### 2021-07-18 ranocha/SummationByPartsOperators.jl#108 @kellertuer
[JOSS Review] 
# Review on the JOSS paper on SummationByPartsOperators.jl

The submitted manuscript is about a Julia package and in general well written.
While the text is easy to read, I am missing a few points from the JOSS Guidelines which should be covered in the paper, I think. I will continuously number my points, just for easier reference, though the later remarks are easier to solve.

The package itself
- has a proper license file
- has substantial scholarly effort and is used in academic papers of the submitting author and his coauthors
- has guidelines how to contribute (a `CONTRIBUTE.md`)
- has a reasonable test coverage of 92%

## Remarks
1. An installation instruction is missing. This can be easily resolved when taking into account the second point
2. I am missing a simple example, i.e. how to get started, solve an easy task with the package. This could also start with how to install the package. It would be nice to have a little more theory with such an example than just the code lines, e.g. as they are given in the “Basic Example” section in the `Readme.md`. Maybe it is not reasonable to put code into the paper (though the examples from the readme also have to be copied line-by-line currently), but maybe then a section on the mathematical background would be neat? On the other hand for example [here](https://joss.theoj.org/papers/10.21105/joss.02520) also some code was included, so I think that would be neat, too.
3. I noticed that the Summary/Abstract is a one-to-one copy of the first paragraph from the package readme. I am new to reviewing for JOSS, so I am not sure whether that is usual.
4. Concerning related sopftware I have two remarks
  * It does not get clear to what extend there is an overlap of the current package to the mentioned ones. For example it sounds a little bit like the current package can be used to solve _all_ problems from the mentioned packages? A little more detail on the relations would be neat
  * While for the first related software both the title/link to the software are mentioned as well as a reference, oáll others are just mentioned with a reference. Could these maybe be extended to also cover software package names and links to the repositories?

## Minor Remarks
5. line 19: The word “resulting” seems to be mispaceed. Just the development of software will most probably not lead to disjoint communities. Maybe you meant that disjoint communities yuield different implementations and what is missing is a joint approach, also because the disjoint communities use different vocabulary?
6. line 32/33: You calim that the package is “reasonably fast”, but this is nowhere supported by a benchmark, comparison or reference. Where does this claim originate from? Sure there is a Benchmark section in the documentation, but I am missing a comparison to other approaches/packages for such a claim. This might also be, because I am – concerning these numerics – not an expert on performance. It would be nice to still justify this claim.
7. line 35: The claim that the software was used in graduate courses is also without reference/example nor what the benefits of using it were.
8. line 55: the term “both” is followed by an enumeration of three terms (with the correct use of a comma before the “and”), so “both” is a wrong reference for this
9. Conerning the API, the available functions are currently listed alphabetically on one page. Might it be a reasonable point to organise them a little more by topic?

## Typographical Remarks

10. line 30: Should the package name maybe be set as code, i.e. not SummationByParts.jl but `SummationByParts.jl`?
11. line 43: The word “basically” can maybe be left out?
12. lines 49/50: the words “artificial dissipation” is used in line 49 and again directly in the first example line 50, so this is a repetition. Can this maybe be avoided?
13. the example from the readme, especially the command `D = periodic_derivative_operator(derivative_order=1, accuracy_order=2, xmin=0.0, xmax=2.0, N=20)` should return a `show` which is a constructor instead of an explaining text and a reference in REPL. I looked a while for a reference for this claim and can not find it, but I am used to `show` methods to print a constructor that recreates the type.
14.   Also more like two short techical remarks: The “Edit on GitHub” links from the Documentation for Contributing and License do not work. I think this might be due to the reason that they are included atuomatically from the main repository (which is of course very good to avoid duplication). Can you maybe check how this might be avoidable?

While this is a little longish, let me summarise that in general it is of course still a good submission and should be accepted after these points are addressed :)
