### 2021-07-18 #3454 @kellertuer
I wrote a few points to review [here](https://github.com/ranocha/SummationByPartsOperators.jl/issues/108) which we surely can also discuss there and I will check the remaining points atop as soon as they are done.
