### 2021-07-19 9f20af0
revisions based on #108
