+++ paper.bib
@@ -319,7 +319,7 @@
 }
 
 @article{ramadhan2020oceananigans,
-  title={Oceananigans. jl: {F}ast and friendly geophysical fluid dynamics
+  title={Oceananigans.jl: {F}ast and friendly geophysical fluid dynamics
          on GPUs},
   author={Ramadhan, Ali and Wagner, Gregory LeClaire and Hill, Chris and
           Campin, Jean-Michel and Churavy, Valentin and Besard, Tim and
