+++ paper.md
@@ -16,7 +16,7 @@ authors:
 affiliations:
  - name: Applied Mathematics Münster, University of Münster, Germany
    index: 1
-date: 25 May 2021
+date: 19 July 2021
 bibliography: paper.bib
 ---
 
@@ -41,8 +41,12 @@ sacrificing flexibility.
 
 Partial differential equations (PDEs) are widely used in science and engineering
 to create mathematical models of real-world processes. Since PDEs often need to
-be solved numerically, a vast amount of numerical methods has been developed,
-resulting sometimes in completely disjoint communities. To transfer developments
+be solved numerically, a vast amount of numerical methods has been developed.
+Since it is impossible to keep up with all recent research, sub-communities
+focussing on specific applications or numerical methods emerged. Sometimes,
+these communities develop different vocabulary and notations, making it hard for
+newcomers (or even experienced researchers) to see similarities and connections
+between seemingly unrelated approaches. To transfer new ideas and developments
 and knowledge from one community to another, common abstractions can be helpful.
 The concept of SBP operators is such an abstraction.
 In recent years, SBP operators have attracted a lot of attention, in particular
@@ -61,7 +65,10 @@ provides a unified interface for different SBP operators. At the same time,
 the implementations are reasonably fast. Together, this facilitates the development
 of new algorithms and research in numerical analysis, which is the primary goal
 of this package. In addition, SummationByPartsOperators.jl has been used in a
-number of graduate-level numerical analysis courses.
+number of graduate-level numerical analysis courses, allowing students to
+understand the connections between different SBP methods by presenting them in
+a unified framework. In addition, some of the operators were not available in
+open source software previously (to the author's knowledge).
 
 
 # Features
@@ -74,7 +81,7 @@ of the following classes:
 - continuous Galerkin methods
 - discontinuous Galerkin methods
 
-Since a discrete derivative operator is basically a linear operator, all of
+Since a discrete derivative operator is a linear operator, all of
 these SBP operators implement the basic interface of such linear operators
 (`AbstractMatrix` in Julia) such as multiplication by vectors and addition of
 operators. Finite difference and Fourier operators on periodic domains also
@@ -82,29 +89,48 @@ allow the construction of rational functions of operators and their efficient
 implementation using the fast Fourier transform [@frigo2005design].
 
 In addition to basic SBP derivative operators, SummationByPartsOperators.jl
-contains a number of artificial dissipation and filtering operators, such as
+contains a number of related operators, such as
 
 - SBP artificial dissipation operators
 - spectral viscosity operators for Fourier methods
 - modal filter operators for Fourier methods and Legendre pseudospectral methods
 
+Using Julia's rich type system, all of these operators are implemented as their
+own types. This enables several optimizations such as a memory requirement
+independent of the number of grid points. In contrast, implementations based
+on sparse/banded matrices have a memory requirement growing linearly with the
+number of grid points. In addition, the structure of the operators can be taken
+into account for operator-vector multiplications, usually resulting in speed-ups
+of an order of magnitude or more on consumer hardware. For example, the application
+of the sixth-order (in the interior) finite difference SBP operator on a grid
+with 1000 nodes takes roughly 330 ns on a consumer CPU from 2017 (Intel® Core™ i7-8700K)
+using version v0.5.5 of SummationByPartsOperators.jl. In contrast, the same
+operation takes roughly 3.9 microseconds using a sparse matrix format used in
+other implementations of this operator [@almquist2017optimized].
+
 Following good software development practices, SummationByPartsOperators.jl
[…]
