### 2021-07-19 ranocha/SummationByPartsOperators.jl#108 @ranocha
Thank you very much for your prompt and thorough review, @kellertuer! Please find below some comments on the issues raised above. I created PR #109 to resolve your remaining comments. I would be happy if you could review #109 and discuss any additional/remaining issues ehre or in that PR. I hope we can resolve this issue soon.

- > An installation instruction is missing. This can be easily resolved when taking into account the second point

  As far as I have understood the [review criteria of JOSS](https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions), an installation instruction should be included in the documentation of the software, not in the paper itself. I expect the reason for this is that the installation procedure might change while the software evolves while the paper is static after publication.
  Nevertheless, I added a sentence mentioning that SummationByPartsOperators.jl is a registered Julia package to the paper. If you think that this is insufficient and a more detailed Julia code example should be included in the paper, please tell me and I will be happy to prepare such an example.
- > I am missing a simple example, i.e. how to get started, solve an easy task with the package. This could also start with how to install the package. It would be nice to have a little more theory with such an example than just the code lines, e.g. as they are given in the “Basic Example” section in the Readme.md. Maybe it is not reasonable to put code into the paper (though the examples from the readme also have to be copied line-by-line currently), but maybe then a section on the mathematical background would be neat? On the other hand for example [here](https://joss.theoj.org/papers/10.21105/joss.02520) also some code was included, so I think that would be neat, too.

  I tried to follow the [JOSS guideline for authors](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain), stating that
  > software documentation such as API (Application Programming Interface) functionality should not be in the paper and instead should be outlined in the software documentation

  Thus, it appears inappropriate to include such an example in the paper itself. I guess the reasoning is that the software will evolve but the paper is static. Thus, it should be avoided to document an API that might be out of date in the future.
- > I noticed that the Summary/Abstract is a one-to-one copy of the first paragraph from the package readme. I am new to reviewing for JOSS, so I am not sure whether that is usual.

  Given that JOSS is an open access software journal and the paper [should contain](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)

  > a summary describing the high-level functionality and purpose of the software

  this paragraph appeared to do the job. If you think I should modify it, please let me know.
- > line 30: Should the package name maybe be set as code, i.e. not SummationByParts.jl but `SummationByParts.jl`?

  I don't think so since it is the name of a package, not some real (Julia) code. The example [you have linked above](https://joss.theoj.org/papers/10.21105/joss.02520) does also not use monospaced font for the package name.
- > the example from the readme, especially the command `D = periodic_derivative_operator(derivative_order=1, accuracy_order=2, xmin=0.0, xmax=2.0, N=20)` should return a show which is a constructor instead of an explaining text and a reference in REPL. I looked a while for a reference for this claim and can not find it, but I am used to show methods to print a constructor that recreates the type.

  I would be happy to see a reference for this request. In my experience, such a behavior is not generic in Julia. For example, the SciML ecosystem follows the approach to provide high-level information about objects instead of (used also by SummationByPartsOperators.jl):
  ```julia
  julia> using OrdinaryDiffEq

  julia> ode = ODEProblem((u, p, t) -> u, 1.0, (0.0, 1.0))
  ODEProblem with uType Float64 and tType Float64. In-place: false
  timespan: (0.0, 1.0)
  u0: 1.0

  julia> sol = solve(ode, Tsit5())
  retcode: Success
  Interpolation: specialized 4th order "free" interpolation
  t: 5-element Vector{Float64}:
  0.0
  0.10001999200479662
  0.3480203186480749
  0.6830525653177172
  1.0
  u: 5-element Vector{Float64}:
  1.0
  1.105193012902056
  1.4162610240705922
  1.9799122739515003
  2.718281708773342
  ```
  In addtion, other basic pacakges do also not print something that can be used directly as constructor, e.g.
  ```julia
  julia> x = SVector(1.0, 2.0)
  2-element SVector{2, Float64} with indices SOneTo(2):
  1.0
  2.0
  ```
  Of course, something like `x = range(0.0, 1.0, length=51); plot(x, sinpi.(x))` will also not print a constructor-like information.
- > Also more like two short techical remarks: The “Edit on GitHub” links from the Documentation for Contributing and License do not work. I think this might be due to the reason that they are included atuomatically from the main repository (which is of course very good to avoid duplication). Can you maybe check how this might be avoidable?

  I couldn't find a way to do so and [filed an issue](https://github.com/JuliaDocs/Documenter.jl/issues/1650).
- > Conerning the API, the available functions are currently listed alphabetically on one page. Might it be a reasonable point to organise them a little more by topic?

  Such an API documentation sorted by topic is basically provided in the section [Introduction](https://ranocha.de/SummationByPartsOperators.jl/stable/introduction/) in the docs. Additional example-guided API documentation is provided in form of tutorials such as [on the linear advection equation](https://ranocha.de/SummationByPartsOperators.jl/stable/tutorials/constant_linear_advection/).
