### 2021-07-19 ranocha/SummationByPartsOperators.jl#108 @kellertuer
For the first point I am fine with mentioning that it is a registered package; that solves point 1.

Concerning 2) For the example, we seem to have different views. Let me carefully elaborate on mine.
I think an easy to start example would be helpful to a reader of the paper. It does not have to be the most elaborate one or the most complicated – mainly to introduce the main mathematical background and directly illustrate how that maps to some code. Note that some have even [quite long](https://joss.theoj.org/papers/10.21105/joss.02520) examples – where one might argue that “printing” so much code might not be what one wants.

The argument with the API seems to me bad in both views that one might have to that: (1) if your API does changes so much as you claim, the package might not be as mature as it should be for a paper? (2) An example could also use only those (basic/main) API parts that are mature enough so the example would also work in the (far) future. Finally the example could mention the version of the package to be reproducible for sure.

Concerning 3) I am sure content wise it summarises the package well – I am not sure whether it is meant to be that you copy that fro somewhere (in harsh words one could think about self-plagiarism?)

Concerning 4) a) seems to be resolved and b) I addressed in #109 with a small comment

Points 5--13 are addressed in #109 I think; and cool that you fixed 14.

So only 2&3 are left to discuss here.
