### 2021-07-19 ranocha/SummationByPartsOperators.jl#108 @kellertuer
A I forgot to talk about the show method – thanks for the details, then you show method might also be fine. As I said I could not find the reference for the show methods we do, but maybe that also depends on the simplicity of the type (struct) at hand; usually our structs are quite plain ones.
