### 2021-07-19 2ed9cb9
Apply suggestions from code review
Co-authored-by: Ronny Bergmann <git@ronnybergmann.net>
