### 2021-07-19 a93d2d7
add links to repos of related software
