### 2021-07-19 503c19b
link Zenodo DOIs of related packages wherever possible
