### 2021-07-19 5c888b7
Revert "link Zenodo DOIs of related packages wherever possible"
This reverts commit d2962f3f0406510d82b67f29a7e12935a828bf9e.
