### 2021-07-19 ebf7cf3
add remark on specialized implementations and multiple dispatch
