### 2021-07-19 ranocha/SummationByPartsOperators.jl#108 @ranocha
Concerning 2): I added such a basic example in c663429 in #109.
