### 2021-07-19 ranocha/SummationByPartsOperators.jl#108 @ranocha
Concerning 3): I updated the summary of the package in the README.md.
