### 2021-07-19 ranocha/SummationByPartsOperators.jl#108 @kellertuer
Cool. Then up to the minor remarks over in #109 we are done and I just have to check the final PDF again (and my checkpoints I ddi not check over in the JOSS review issue), but this Issue can be already closed I think.
