### 2021-07-19 231f4d7
Update paper.md
Co-authored-by: Ronny Bergmann <git@ronnybergmann.net>
