### 2021-07-20 #3454 @kellertuer
I had a nice discussion with @ranocha yesterday in the linked issue I opened in the repository of SummationByPartsOperators.jl (and in the corresponding PR) and I checked all points in the check list above, since I after these changes I can propose to accept the paper.
