### 2021-07-22 8b6c318
update reference xiao2021kinetic
