+++ paper.bib
@@ -333,12 +333,16 @@
   doi={10.21105/joss.02018}
 }
 
-@misc{xiao2021kinetic,
-  title={Kinetic.jl: {A} portable finite volume toolbox in {J}ulia},
+@article{xiao2021kinetic,
+  title={Kinetic.jl: {A} portable finite volume toolbox for scientific and
+         neural computing},
   author={Xiao, Tianbai},
+  journal={Journal of Open Source Software},
+  volume={6},
+  number={62},
+  pages={3060},
   year={2021},
-  howpublished={\url{https://github.com/vavrines/Kinetic.jl}},
-  note={In review in Journal of Open Source Software}
+  doi={10.21105/joss.03060}
 }
 
 @misc{oreilly2019sbp,
