### 2021-08-04 7f15656
set development version to v0.5.7-pre
