### 2021-08-20 #3454 @dawbarton
Sorry for the delay; I've now been through the paper and the package fairly thoroughly. Overall, this looks like a good contribution to JOSS with substantive contributions to the community and well written documentation that covers most bases. The paper has even motivated me to look into the broader SBP literature (I sit in one of the sub-communities)!

A few very minor points that you can feel free to ignore:

1. In the README.md: `If you want to update SummationByPartsOperators.jl, you can use` while this is true, it will also update *all* other packages in your current project, which might not be desirable, particularly if you use fast moving parts of the ecosystem. (Of course separate projects is the gold standard but many people don't use them.)
2. In the documentation on `constant_linear_advection`: you use the SBP operator directly as the parameter (since it is the only parameter), I'd suggest replacing it with a named tuple as you do in some of the later examples. I've had issues recently of some of my students blindly copying code from examples and not really understanding how parameter passing can be done and making a (surprisingly large) mess of it. Using a named tuple from the start might just give people unfamiliar with Julia a nudge in the right direction. (Given the context, this is personal preference so feel free to ignore.)
3. The benchmarks webpage is very dense, to the point that I'm not sure what I'm supposed to be taking from it. While I like the new `@benchmark` output for individual tests, to have so many of them close together is overwhelming. Just a simple `@btime` with a bit of covering text to interpret the results might be more helpful to people looking through the results. (I'd have to spend a reasonable amount of time trying to understand them, to the point where it might be easier for me to cut and paste the code into Julia myself to see what is going on.)

Overall, this looks good to go from my perspective (save the optional comments above).
