### 2021-08-20 4eb9f05
clarify update command
