+++ docs/src/index.md
@@ -27,6 +27,16 @@ is a registered Julia package. Thus, you can install it from the Julia REPL via
 julia> using Pkg; Pkg.add("SummationByPartsOperators")
 ```
 
+If you want to update SummationByPartsOperators.jl, you can use
+```julia
+julia> using Pkg; Pkg.update("SummationByPartsOperators")
+```
+As usual, if you want to update SummationByPartsOperators.jl and all other
+packages in your current project, you can execute
+```julia
+julia> using Pkg; Pkg.update()
+```
+
 
 ## Basic examples
 
