### 2021-08-20 61f5de8
use a named tuple for parameters in the linear advection tutorial
