+++ docs/src/benchmarks.md
@@ -6,8 +6,11 @@ on virtual machines in the cloud to generate the documentation automatically.
 
 ## First-derivative operators
 
-Periodic domains:
-```@example
+#### Periodic domains
+
+Let's set up some benchmark code.
+
+```@example first-derivative-periodic
 using BenchmarkTools
 using LinearAlgebra, SparseArrays
 using SummationByPartsOperators, DiffEqOperators
@@ -34,15 +37,32 @@ function doit(D, text, du, u)
   show(stdout, MIME"text/plain"(), @benchmark mul!($du, $D, $u))
   println()
 end
+```
 
+First, we benchmark the implementation from SummationByPartsOperators.jl.
+```@example first-derivative-periodic
 doit(D_SBP, "D_SBP:", du, u)
-doit(D_DEO, "D_DEO:", du, u)
+```
+
+Next, we compare this to the runtime obtained using a sparse matrix representation
+of the derivative operator. Depending on the hardware etc., this can be an order
+of magnitude slower than the optimized implementation from SummationByPartsOperators.jl.
+```@example first-derivative-periodic
 doit(D_sparse, "D_sparse:", du, u)
 ```
 
+Finally, we benchmark the implementation of the same derivative operator in
+DiffEqOperators.jl.
+```@example first-derivative-periodic
+doit(D_DEO, "D_DEO:", du, u)
+```
 
-Bounded domains:
-```@example
+
+#### Bounded domains
+
+We start again by setting up some benchmark code.
+
+```@example first-derivative-bounded
 using BenchmarkTools
 using LinearAlgebra, SparseArrays
 using SummationByPartsOperators, BandedMatrices
@@ -66,16 +86,32 @@ function doit(D, text, du, u)
   show(stdout, MIME"text/plain"(), @benchmark mul!($du, $D, $u))
   println()
 end
+```
 
+First, we benchmark the implementation from SummationByPartsOperators.jl.
+```@example first-derivative-bounded
 doit(D_SBP, "D_SBP:", du, u)
+```
+
+Again, we compare this to a representation of the derivative operator as a
+sparse matrix. No surprise - it is again much slower, as in periodic domains.
+```@example first-derivative-bounded
 doit(D_sparse, "D_sparse:", du, u)
+```
+
+FInally, we compare it to a representation as banded matrix. Disappointingly,
+this is still much slower than the optimized implementation from
+SummationByPartsOperators.jl.
+```@example first-derivative-bounded
 doit(D_banded, "D_banded:", du, u)
 ```
 
 
 ## Dissipation operators
 
-```@example
+We follow the same structure as before. At first, we set up some benchmark code.
+
+```@example dissipation
 using BenchmarkTools
 using LinearAlgebra, SparseArrays
 using SummationByPartsOperators, BandedMatrices
@@ -101,11 +137,26 @@ function doit(D, text, du, u)
   show(stdout, MIME"text/plain"(), @benchmark mul!($du, $D, $u))
   println()
 end
+```
 
+At first, let us benchmark the derivative and dissipation operators implemented
+in SummationByPartsOperators.jl.
+```@example dissipation
 doit(D_SBP, "D_SBP:", du, u)
 doit(Di_SBP, "Di_SBP:", du, u)
+```
+
+Next, we compare the results to sparse matrix representations. It will not
+come as a surprise that these are again much (around an order of magnitude)
+slower.
+```@example dissipation
 doit(Di_sparse, "Di_sparse:", du, u)
 doit(Di_banded, "Di_banded:", du, u)
+```
+
+Finally, let's benchmark the same computation if a full (dense) matrix is used
[…]
