### 2021-08-20 d434a54
set development version to v0.5.8-pre
