### 2021-08-20 #3454 @ranocha
Thank you very much for your review, @dawbarton.

1. I updated this part of the README.md in https://github.com/ranocha/SummationByPartsOperators.jl/commit/4eb9f057f179e510656a6f56b7a42f3ee5c4779c.
2. I followed your good suggestion and made the corresponding changes in https://github.com/ranocha/SummationByPartsOperators.jl/commit/61f5de83386819bb5690c1cff91aac0c0a78bc43.
3. Yeah, this page is particularly dense, I agree. It was mainly aimed at advanced people and not average users. Since I do like the additional information on the distribution of the results returned by `@benchmark`, I would rather like to avoid switching to only `@btime`. However, I tried to make this page less dense by adding additional text in https://github.com/ranocha/SummationByPartsOperators.jl/commit/ec0ad0358ed43016bd71e0c9c0e89e85f809d290, explaining some steps and results a bit more instead of just presenting the numbers.

The updated version of the documentation is available in the [dev docs](https://ranocha.de/SummationByPartsOperators.jl/dev/benchmarks/); a new version should be registered soon.
