### 2021-08-20 #3454 @dawbarton
> Is there anything else about the paper itself? I'm asking since you didn't check any of the boxes in your section "Software paper" above.

Sorry that was me doing things in several parts and then forgetting to check the outstanding boxes. No, nothing on the paper, it looks good.
