+++ paper.bib
@@ -178,15 +178,18 @@
 }
 
 
-@online{ranocha2020preventing,
+@article{ranocha2021preventing,
   title={Preventing pressure oscillations does not fix local linear stability
          issues of entropy-based split-form high-order schemes},
   author={Ranocha, Hendrik and Gassner, Gregor J},
-  year={2020},
-  month={09},
+  journal={Communications on Applied Mathematics and Computation},
+  year={2021},
+  month={08},
+  doi={10.1007/s42967-021-00148-z},
   eprint={2009.13139},
-  eprinttype={arXiv},
-  eprintclass={math.NA}
+  eprinttype={arxiv},
+  eprintclass={math.NA},
+  keywords={journal}
 }
 @misc{ranocha2020preventing-repro,
   title={Reproducibility: {P}reventing pressure oscillations does not fix local
