### 2021-08-24 #3454 @VivianePons
Hi all, I am glad to see we are almost done! Thank you @dawbarton & @kellertuer for the reviews and @ranocha for the improvements on the software.

We can now move towards acceptance, @ranocha can you make a tagged release of the software? (Zenodo or Figshare for example). Please be careful to use the same metadata as the paper (title and authors). Send me the DOI when done. Thanks
