### 2021-08-24 #3454 @ranocha
Version v0.5.7 (including all revisions based on the reviewers' comments) of SummationByPartsOperators.jl is registered with [DOI 10.5281/zenodo.5226913](https://doi.org/10.5281/zenodo.5226913) at Zenodo. Please let me know whether that's okay or whether I need to update anything.
